+++ Net_Core/Web.Support/BaseControllers/AbstractGroupController.cs
@@ -130,7 +130,7 @@ public abstract class AbstractGroupController : ControllerBase
         {
             Tuple<string, string> mOrderByAndWhere = SearchUtility.GetOrderByAndWhere(mColumns, searchCriteria.searchColumns, searchCriteria.sortColumns, searchCriteria.searchText);
             string mOrderByClause = mOrderByAndWhere.Item1;
-            string mWhereClause = mOrderByAndWhere.Item2;
+            string mWhereClause = mOrderByAndWhere.Item2 + " AND Security_Entity_SeqID = " + SecurityEntityUtility.CurrentProfile.Id.ToString();
             MSearchCriteria mSearchCriteria = new MSearchCriteria
             {
                 Columns = mColumns,